requirements.txt
youtube_dl
youtube-search-python
discord.py
PyNaCl
ffmpeg
numpy==1.22.1
python-dotenv